requirements.txt
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.0.0
python-dotenv==1.1.1
